Refactor base module into affine and quadratic submodules
